requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2
python-multipart==0.0.6
ydata-profiling==4.5.1
chromadb==0.4.17
llama-cpp-python==0.2.23
langchain==0.0.350
transformers==4.35.2
python-dotenv==1.0.0
torch==2.1.2
tqdm==4.66.1
sentence-transformers==2.2.2
